Move encode_fp Python function from transform.py to maker.py
The encode_fp function uses the Python implementation of data fingerprints.
To use the Python implementation of data fingerprints, one must build the
Docker image and execute within the Docker container, otherwise, the data
fingerprints Python package is not available. Since encode_fp was located
in transform.py, any code that makes use of tranform.py, even if it does
not use the Python implementation of data fingerprints, will fail.
Therefore, move encode_fp function to maker.py so that code which makes use
of functions defined in transform.py can be run outside of the Docker
container. Note that, even with this change, hello_docket.py still can only
be executed within the Docker container since it uses encode_fp. However,
hello_docket.nf was revised to use the Perl implementation of data finger-
prints rather than the Python version, so it can be run outside of the
Docker container (as long as other dependencies are satisfied). With this
change, the docket_cluster.nf script can also now be run outside of the
Docker container.
